requirements.txt
numpy==1.14.4
Pillow==5.1.0
protobuf==3.6.0
six==1.11.0
tensorboardX==1.2
torch==1.1.0
torchvision==0.3.0
tqdm==4.23.4
